Convert $timeoutTimestamp to $timeoutDuration.
The timestamp value almost always is set as time() + duration, so
exposing it as a duration makes better sense.  Plus, this allows the
value to be given as configuration for the Locker without causing a
long-running Locker to eat into its timeout time (each call to lock
still has its own timeout).